+++ tests/creature.rs
@@ -533,7 +533,7 @@ fn remove_creature_on_alter() {
         type StatisticsSeed = ();
         type StatisticsAlteration = ();
 
-        fn alter(
+        fn alter_statistics(
             &self,
             _character: &mut dyn Character<R>,
             _alteration: &Self::StatisticsAlteration,