+++ tests/event.rs
@@ -148,7 +148,7 @@ fn conditional() {
             Box::new(v.into_iter())
         }
 
-        fn alter(
+        fn alter_abilities(
             &self,
             actor: &mut dyn Actor<CustomRules>,
             alteration: &Self::AbilitiesAlteration,